refine logger, remove all commented-out code, remove DEFAULT_OPTIONS, and restructure folders into more scalable setup
